Split form_fields tests into different files.
