+++ tests/forms_tests/field_tests/test_charfield.py
@@ -1,0 +1,109 @@
+from __future__ import unicode_literals
+
+from django.forms import (
+    CharField, PasswordInput, Textarea, TextInput, ValidationError,
+)
+from django.test import SimpleTestCase
+
+from . import FormFieldAssertionsMixin
+
+
+class CharFieldTest(FormFieldAssertionsMixin, SimpleTestCase):
+
+    def test_charfield_1(self):
+        f = CharField()
+        self.assertEqual('1', f.clean(1))
+        self.assertEqual('hello', f.clean('hello'))
+        with self.assertRaisesMessage(ValidationError, "'This field is required.'"):
+            f.clean(None)
+        with self.assertRaisesMessage(ValidationError, "'This field is required.'"):
+            f.clean('')
+        self.assertEqual('[1, 2, 3]', f.clean([1, 2, 3]))
+        self.assertIsNone(f.max_length)
+        self.assertIsNone(f.min_length)
+
+    def test_charfield_2(self):
+        f = CharField(required=False)
+        self.assertEqual('1', f.clean(1))
+        self.assertEqual('hello', f.clean('hello'))
+        self.assertEqual('', f.clean(None))
+        self.assertEqual('', f.clean(''))
+        self.assertEqual('[1, 2, 3]', f.clean([1, 2, 3]))
+        self.assertIsNone(f.max_length)
+        self.assertIsNone(f.min_length)
+
+    def test_charfield_3(self):
+        f = CharField(max_length=10, required=False)
+        self.assertEqual('12345', f.clean('12345'))
+        self.assertEqual('1234567890', f.clean('1234567890'))
+        msg = "'Ensure this value has at most 10 characters (it has 11).'"
+        with self.assertRaisesMessage(ValidationError, msg):
+            f.clean('1234567890a')
+        self.assertEqual(f.max_length, 10)
+        self.assertIsNone(f.min_length)
+
+    def test_charfield_4(self):
+        f = CharField(min_length=10, required=False)
+        self.assertEqual('', f.clean(''))
+        msg = "'Ensure this value has at least 10 characters (it has 5).'"
+        with self.assertRaisesMessage(ValidationError, msg):
+            f.clean('12345')
+        self.assertEqual('1234567890', f.clean('1234567890'))
+        self.assertEqual('1234567890a', f.clean('1234567890a'))
+        self.assertIsNone(f.max_length)
+        self.assertEqual(f.min_length, 10)
+
+    def test_charfield_5(self):
+        f = CharField(min_length=10, required=True)
+        with self.assertRaisesMessage(ValidationError, "'This field is required.'"):
+            f.clean('')
+        msg = "'Ensure this value has at least 10 characters (it has 5).'"
+        with self.assertRaisesMessage(ValidationError, msg):
+            f.clean('12345')
+        self.assertEqual('1234567890', f.clean('1234567890'))
+        self.assertEqual('1234567890a', f.clean('1234567890a'))
+        self.assertIsNone(f.max_length)
+        self.assertEqual(f.min_length, 10)
+
+    def test_charfield_length_not_int(self):
+        """
+        Setting min_length or max_length to something that is not a
+        number returns an exception.
+        """
+        with self.assertRaises(ValueError):
+            CharField(min_length='a')
+        with self.assertRaises(ValueError):
+            CharField(max_length='a')
+        with self.assertRaises(ValueError):
+            CharField('a')
+
+    def test_charfield_widget_attrs(self):
+        """
+        CharField.widget_attrs() always returns a dictionary (#15912).
+        """
+        # Return an empty dictionary if max_length is None
+        f = CharField()
+        self.assertEqual(f.widget_attrs(TextInput()), {})
+        self.assertEqual(f.widget_attrs(Textarea()), {})
+
+        # Otherwise, return a maxlength attribute equal to max_length
+        f = CharField(max_length=10)
+        self.assertEqual(f.widget_attrs(TextInput()), {'maxlength': '10'})
+        self.assertEqual(f.widget_attrs(PasswordInput()), {'maxlength': '10'})
+        self.assertEqual(f.widget_attrs(Textarea()), {'maxlength': '10'})
+
+    def test_charfield_strip(self):
+        """
+        Values have whitespace stripped and that strip=False works.
+        """
+        f = CharField()
+        self.assertEqual(f.clean(' 1'), '1')
+        self.assertEqual(f.clean('1 '), '1')
+
+        f = CharField(strip=False)
+        self.assertEqual(f.clean(' 1'), ' 1')
+        self.assertEqual(f.clean('1 '), '1 ')
+
+    def test_charfield_disabled(self):
+        f = CharField(disabled=True)
+        self.assertWidgetRendersTo(f, '<input type="text" name="f" id="id_f" disabled />')
