+++ tests/forms_tests/field_tests/test_durationfield.py
@@ -1,0 +1,42 @@
+from __future__ import unicode_literals
+
+import datetime
+
+from django.forms import DurationField, Form
+from django.test import SimpleTestCase
+from django.utils.duration import duration_string
+
+from . import FormFieldAssertionsMixin
+
+
+class DurationFieldTest(FormFieldAssertionsMixin, SimpleTestCase):
+
+    def test_durationfield_1(self):
+        f = DurationField()
+        self.assertEqual(datetime.timedelta(seconds=30), f.clean('30'))
+        self.assertEqual(datetime.timedelta(minutes=15, seconds=30), f.clean('15:30'))
+        self.assertEqual(datetime.timedelta(hours=1, minutes=15, seconds=30), f.clean('1:15:30'))
+        self.assertEqual(
+            datetime.timedelta(days=1, hours=1, minutes=15, seconds=30, milliseconds=300),
+            f.clean('1 1:15:30.3')
+        )
+
+    def test_durationfield_2(self):
+        class DurationForm(Form):
+            duration = DurationField(initial=datetime.timedelta(hours=1))
+        f = DurationForm()
+        self.assertHTMLEqual(
+            '<input id="id_duration" type="text" name="duration" value="01:00:00">',
+            str(f['duration'])
+        )
+
+    def test_durationfield_integer_value(self):
+        f = DurationField()
+        self.assertEqual(datetime.timedelta(0, 10800), f.clean(10800))
+
+    def test_durationfield_prepare_value(self):
+        field = DurationField()
+        td = datetime.timedelta(minutes=15, seconds=30)
+        self.assertEqual(field.prepare_value(td), duration_string(td))
+        self.assertEqual(field.prepare_value('arbitrary'), 'arbitrary')
+        self.assertIsNone(field.prepare_value(None))
