+++ tests/forms_tests/field_tests/test_typedchoicefield.py
@@ -1,0 +1,81 @@
+from __future__ import unicode_literals
+
+import decimal
+
+from django.forms import TypedChoiceField, ValidationError
+from django.test import SimpleTestCase
+from django.utils import six
+
+from . import FormFieldAssertionsMixin
+
+
+class TypedChoiceFieldTest(FormFieldAssertionsMixin, SimpleTestCase):
+
+    def test_typedchoicefield_1(self):
+        f = TypedChoiceField(choices=[(1, "+1"), (-1, "-1")], coerce=int)
+        self.assertEqual(1, f.clean('1'))
+        msg = "'Select a valid choice. 2 is not one of the available choices.'"
+        with self.assertRaisesMessage(ValidationError, msg):
+            f.clean('2')
+
+    def test_typedchoicefield_2(self):
+        # Different coercion, same validation.
+        f = TypedChoiceField(choices=[(1, "+1"), (-1, "-1")], coerce=float)
+        self.assertEqual(1.0, f.clean('1'))
+
+    def test_typedchoicefield_3(self):
+        # This can also cause weirdness: be careful (bool(-1) == True, remember)
+        f = TypedChoiceField(choices=[(1, "+1"), (-1, "-1")], coerce=bool)
+        self.assertTrue(f.clean('-1'))
+
+    def test_typedchoicefield_4(self):
+        # Even more weirdness: if you have a valid choice but your coercion function
+        # can't coerce, you'll still get a validation error. Don't do this!
+        f = TypedChoiceField(choices=[('A', 'A'), ('B', 'B')], coerce=int)
+        msg = "'Select a valid choice. B is not one of the available choices.'"
+        with self.assertRaisesMessage(ValidationError, msg):
+            f.clean('B')
+        # Required fields require values
+        with self.assertRaisesMessage(ValidationError, "'This field is required.'"):
+            f.clean('')
+
+    def test_typedchoicefield_5(self):
+        # Non-required fields aren't required
+        f = TypedChoiceField(choices=[(1, "+1"), (-1, "-1")], coerce=int, required=False)
+        self.assertEqual('', f.clean(''))
+        # If you want cleaning an empty value to return a different type, tell the field
+
+    def test_typedchoicefield_6(self):
+        f = TypedChoiceField(choices=[(1, "+1"), (-1, "-1")], coerce=int, required=False, empty_value=None)
+        self.assertIsNone(f.clean(''))
+
+    def test_typedchoicefield_has_changed(self):
+        # has_changed should not trigger required validation
+        f = TypedChoiceField(choices=[(1, "+1"), (-1, "-1")], coerce=int, required=True)
+        self.assertFalse(f.has_changed(None, ''))
+        self.assertFalse(f.has_changed(1, '1'))
+        self.assertFalse(f.has_changed('1', '1'))
+
+        f = TypedChoiceField(
+            choices=[('', '---------'), ('a', "a"), ('b', "b")], coerce=six.text_type,
+            required=False, initial=None, empty_value=None,
+        )
+        self.assertFalse(f.has_changed(None, ''))
+        self.assertTrue(f.has_changed('', 'a'))
+        self.assertFalse(f.has_changed('a', 'a'))
+
+    def test_typedchoicefield_special_coerce(self):
+        """
+        Test a coerce function which results in a value not present in choices.
+        Refs #21397.
+        """
+        def coerce_func(val):
+            return decimal.Decimal('1.%s' % val)
+
+        f = TypedChoiceField(choices=[(1, "1"), (2, "2")], coerce=coerce_func, required=True)
+        self.assertEqual(decimal.Decimal('1.2'), f.clean('2'))
+        with self.assertRaisesMessage(ValidationError, "'This field is required.'"):
+            f.clean('')
+        msg = "'Select a valid choice. 3 is not one of the available choices.'"
+        with self.assertRaisesMessage(ValidationError, msg):
+            f.clean('3')
